Add exp file
exp9 uses the data augmentation minimum of 0.2 and traffic_images/Dataset_360 as training.val does not modify

ex10 adjusted the data augmentation minimum to 0.5. Other environments are like ex9

ex11 adjusts the minimum data augmentation to 0.2 and randomly divides the dataset of the existing Dataset_360 into 8:1:1, and then proceeds with learning

ex 12 adjusts the minimum data augmentation to 0.5, and the rest are all the same as ex11.